requirements.txt
python-socketio==5.10.0
python-dotenv==1.0.0
eventlet==0.35.2
Flask==3.0.2
hx711==1.1.2
gpiozero==2.0

